Renamed `generator.js` to be `index.js`
This is so that the main file will  be the `index.js` file. This way we can export it and get the same functionality.
