+++ app/index.js
@@ -1,46 +1,55 @@
-import program from 'commander';
-import generator from './generator';
-import updateNotifier from 'update-notifier';
-import pkg from './../package.json';
+import path from 'path';
+import { prepare } from './input';
+import * as models from './models';
+import * as output from './output';
+import to from 'to-js';
 
-export default function() {
-  // check for update and notify
-  updateNotifier({ pkg }).notify();
-
-  // get the inputs
-  program
-    .version(pkg.version)
-    .usage('fakeit [options]')
-    .option('-o, --output [value]', 'The output format to generate.  Supported formats are: json, csv, yaml, cson', 'json')
-    .option('-a, --archive [value]', 'The archive file to generate.  Supported formats are: zip')
-    .option('-m, --models [value]', 'A directory or comma-delimited list of files models to use.', process.cwd())
-    .option('-d, --destination [value]', 'The output destination.  Values can be: couchbase, console or a directory path.', 'console')
-    .option('-f, --format [value]', 'The spacing format to use for JSON and YAML file generation.  Default is 2', 2)
-    .option('-n, --number [value]', 'Overrides the number of documents to generate specified by the model.')
-    .option('-i, --input [value]', 'Directory or comma-delimited list of files to use as inputs.  Support formats are: json, yaml, csv, cson, zip')
-    .option('-s, --server [address]', 'Couchbase Server or Sync-Gateway address', '127.0.0.1')
-    .option('-b, --bucket [name]', 'The name of a Couchbase Bucket.  The default value is: default', 'default')
-    .option('-p, --password [value]', 'Bucket password')
-    .option('-t, --timeout [value]', 'A timeout value for database operations', 5000)
-    .option('-l, --limit [value]', 'Limit the number of save operations at a time.  Default: 100', 100)
-    .option('-u, --username [name]', 'The sync-gateway username')
-    .option('-e, --exclude [model]', 'A comma-delimited list of model names to exclude from output', '')
-    .option('-v, --verbose', 'Whether or not to use verbose output')
-    .parse(process.argv);
-
-  // run the program
-  generator(program)
-    .then(() => {
-      // console.log('Data Generation Complete');
-      process.exit();
-    })
-    .catch((err) => {
-      console.error(err);
-      process.exit(1);
-    });
+export default function fakeit(options = {}) {
+  options = to.extend({
+    output: 'console',
+    archive: '',
+    models: process.cwd(),
+    destination: 'console',
+    format: 2,
+    server: '127.0.0.1',
+    bucket: 'default'
+  }, options);
+  return new Promise((resolve, reject) => {
+    // console.log('generator.start');
+    validate(options);
+    prepare(options)
+      .then(() => models.prepare(options))
+      .then((model_documents_count) => output.prepare(options, resolve, reject, model_documents_count))
+      .then(() => models.generate(options))
+      .catch((err) => {
+        try {
+          output.errorCleanup();
+        } finally {
+          reject(err);
+        }
+      });
+  });
 }
 
-process.on('uncaughtException', (err) => {
-  console.error('An uncaughtException was found:', err.stack);
-  process.exit(1);
-});
+function validate(options) {
+  if ('json,cson,csv,yml,yaml'.indexOf(options.output) === -1) { // validate output format
+    throw new Error('Unsupported output type');
+  } else if (
+    options.archive &&
+    path.extname(options.archive) !== '.zip'
+  ) { // validate archive format
+    throw new Error('The archive must be a zip file');
+  } else if (
+    options.destination === 'couchbase' && (
+      !options.server ||
+      !options.bucket
+    )
+  ) { // validate couchbase
+    throw new Error('For the server and bucket must be specified when outputting to Couchbase');
+  } else if (
+    options.destination === 'couchbase' &&
+    options.archive
+  ) { // validate couchbase
+    throw new Error('The archive option cannot be used when the output is couchbase');
+  }
+}
