+++ package.json
@@ -13,6 +13,7 @@
   "bin": {
     "fakeit": "bin/fakeit"
   },
+  "main": "dist/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/bentonam/fakeit.git"
